Refactor BinaryOpInsn & UnaryOpInsn classes
